Compatible with R4.4.0
2 modifications
t3: `filter(...)[,columns]` returns a data frame with 1 row (because we don't have drop = TRUE). So, sort() is applied to a data.frame. This used to work (with a warning), now errors. New version: explicitly convert t3 to a vector before sorting

tableDEX: input$topM2 is a string (because the UI is a `textInput()`). `head(mtcars, "2")` used to work as `head(mtcars, 2)`, not anymore. Now need to explicitly make numeric.